next: split landing/consts to avoid loading the software fallback data
