+++ src/packages/next/lib/landing/consts.ts
@@ -3,10 +3,6 @@
  *  License: AGPLv3 s.t. "Commons Clause" – see LICENSE.md for details
  */
 
-import { fromPairs } from "lodash";
-
-import { EnvData } from "./types";
-
 export const LANGUAGE_NAMES = [
   "python",
   "R",
@@ -20,21 +16,3 @@ export const LANGUAGE_NAMES = [
 export const SOFTWARE_ENV_NAMES = ["22.04", "20.04", "18.04"] as const;
 export const SOFTWARE_ENV_DEFAULT = "20.04";
 export type SoftwareEnvNames = typeof SOFTWARE_ENV_NAMES[number];
-
-export const SOFTWARE_URLS: { [key in SoftwareEnvNames]: string } = fromPairs(
-  SOFTWARE_ENV_NAMES.map((name) => [
-    name,
-    `https://storage.googleapis.com/cocalc-compute-environment/software-inventory-${name}.json`,
-  ])
-);
-
-import SOFTWARE_1804 from "dist/software-inventory/18.04.json";
-import SOFTWARE_2004 from "dist/software-inventory/20.04.json";
-import SOFTWARE_2204 from "dist/software-inventory/22.04.json";
-
-// Note: we need to be explicit with these rougher types, because TS can't infer them from the JSON files since they're too large.
-export const SOFTWARE_FALLBACK: { [key in SoftwareEnvNames]: EnvData } = {
-  "18.04": SOFTWARE_1804 as EnvData,
-  "20.04": SOFTWARE_2004 as EnvData,
-  "22.04": SOFTWARE_2204 as EnvData,
-} as const;
